+++ platform/psdk_j7/bios/tivx_queue.c
@@ -18,7 +18,7 @@ vx_status tivxQueueCreate(
 {
     vx_status status = (vx_status)VX_FAILURE;
 
-    if ((NULL != queue) && (NULL != queue_memory) && (0 != max_elements))
+    if ((NULL != queue) && (NULL != queue_memory) && (0U != max_elements))
     {
         status = (vx_status)VX_SUCCESS;
 
@@ -107,7 +107,7 @@ vx_status tivxQueuePut(tivx_queue *queue, uintptr_t data, uint32_t timeout)
     do
     {
         /* disable interrupts */
-        cookie = HwiP_disable();
+        cookie = (uint32_t)HwiP_disable();
 
         if (queue->count < queue->max_ele)
         {
@@ -115,7 +115,7 @@ vx_status tivxQueuePut(tivx_queue *queue, uintptr_t data, uint32_t timeout)
             queue->queue[queue->cur_wr] = data;
 
             /* increment put pointer */
-            queue->cur_wr = (queue->cur_wr + 1) % queue->max_ele;
+            queue->cur_wr = (queue->cur_wr + 1U) % queue->max_ele;
 
             /* increment count of number element in que */
             queue->count++;
@@ -199,15 +199,15 @@ vx_status tivxQueueGet(tivx_queue *queue, uintptr_t *data, uint32_t timeout)
     do
     {
         /* disable interrupts */
-        cookie = HwiP_disable();
+        cookie = (uint32_t)HwiP_disable();
 
-        if (queue->count > 0)
+        if (queue->count > 0U)
         {
             /* extract the element */
             *data = queue->queue[queue->cur_rd];
 
             /* increment get pointer */
-            queue->cur_rd = (queue->cur_rd + 1) % queue->max_ele;
+            queue->cur_rd = (queue->cur_rd + 1U) % queue->max_ele;
 
             /* decrmeent number of elements in que */
             queue->count--;
@@ -289,9 +289,9 @@ vx_bool tivxQueueIsEmpty(tivx_queue *queue)
     uint32_t cookie;
 
     /* disable interrupts */
-    cookie = HwiP_disable();
+    cookie = (uint32_t)HwiP_disable();
 
-    if (queue->count == 0)
+    if (queue->count == 0U)
     {
         is_empty = (vx_bool)vx_true_e;
     }
@@ -309,9 +309,9 @@ vx_status tivxQueuePeek(tivx_queue *queue, uintptr_t *data)
     uint32_t cookie;
 
     /* disable interrupts */
-    cookie = HwiP_disable();
+    cookie = (uint32_t)HwiP_disable();
 
-    if (queue->count > 0)
+    if (queue->count > 0U)
     {
         /* 'peek' the element but dont extract it */
         *data = queue->queue[queue->cur_rd];
